pyproject.toml
[tool.poetry]
name = "stdb"
version = "0.1.0"
description = "Python module for creating Synapse External Tables from Databricks"
authors = ["Kolmar Kafran <kolmar@kafran.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.*"
pyspark = "^3.5.1"
azure-identity = "^1.17.1"
databricks-sdk = "^0.29.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.1"
pytest-cov = "^5.0.0"


[tool.poetry.group.doc.dependencies]
mkdocs-material = "^9.5.29"
mkdocstrings = "^0.25.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
